change behavior to just improve the error
